commented loop suggestions
I added a commented-out loop that is part pseudo-code part actual code. It would handle the part about having the water on for 30 seconds and restarting if they take their hand away. It does not handle them having their hands away for 15 seconds but I will add that I also added some comments at the top for comment semantics. And I commented in at the bottom what we need to do to reset.