post loop conditions
I added the conditions for what should be done when the loop exits. I also realized the temp is only displayed before the hands so I made a seperate loop for that. I tried to put some things into the loop and conditions where they would need to go but I don't want to break your portion by doing it wrong so I stopped